change back to using int

+++ src/server.cc
@@ -5,7 +5,7 @@
 
 namespace tl = thallium;
 
-void sum(const tl::request& req, int64_t x, int64_t y, std::vector<int> c, std::vector<int64_t> d, std::vector<int64_t> e) {
+void sum(const tl::request& req, int x, int y, std::vector<int> c, std::vector<int64_t> d, std::vector<int64_t> e) {
     std::cout << "Computing " << x << "+" << y << std::endl;
     req.respond(x+y);
 }
